ml-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
pandas==2.2.3
numpy==2.1.3
pytest==8.3.3
python-dotenv==1.0.1
